requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
google-adk==0.5.0
gitpython==3.1.40
python-dotenv==1.0.1
requests==2.31.0
protobuf==4.25.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
deprecated==1.3.1
wrapt==2.0.1
